+++ projects/framework-flame/resource/bootstrap/init.py
@@ -4,8 +4,7 @@
 import flame
 
 import logging
-import functools
-from ftrack_framework_flame import bootstrap_integration, on_run_tool_callback
+from ftrack_framework_flame import bootstrap_integration, on_run_tool_callback, get_ftrack_menu
 
 from ftrack_utils.extensions.environment import (
     get_extensions_path_from_environment,
@@ -53,67 +52,3 @@ def get_timeline_custom_ui_actions():
 #     # Adds custom actions to the contextual menu available in Batch.
 #     return get_ftrack_menu()
 
-def scope_node(show_on, selection):
-    for item in selection:
-        if isinstance(item, show_on):
-            return True
-
-    return False
-
-
-def get_ftrack_menu(show_on=None):
-    show_on = show_on or ()
-    client_instance, registry_instance = bootstrap_integration(get_extensions_path_from_environment())
-
-    dcc_config = registry_instance.get_one(
-        name='framework-flame', extension_type='dcc_config'
-    )['extension']
-
-    logger.debug(f'Read DCC config: {dcc_config}')
-
-    # Create ftrack menu
-    actions = []
-    # Register tools into ftrack menu
-    for tool in dcc_config['tools']:
-        run_on = tool.get("run_on")
-        on_menu = tool.get("menu", True)
-        if on_menu:
-            new_action = {
-                'name': tool['label'],
-                'execute': functools.partial(
-                    on_run_tool_callback,
-                    client_instance,
-                    tool.get('name'),
-                    tool.get('dialog_name'),
-                    tool['options'],
-                ),
-                'minimumVersion': '2023',
-                'isVisible': functools.partial(
-                    scope_node, show_on
-                ),
-            }
-            actions.append(new_action)
-
-        if run_on:
-            if run_on == "startup":
-                # Execute startup tool-configs
-                on_run_tool_callback(
-                    client_instance,
-                    tool.get('name'),
-                    tool.get('dialog_name'),
-                    tool['options'],
-                )
-            else:
-                logger.error(
-                    f"Unsupported run_on value: {run_on} tool section of the "
-                    f"tool {tool.get('name')} on the tool config file: "
-                    f"{dcc_config['name']}. \n Currently supported values:"
-                    f" [startup]"
-                )
-
-    return [
-        {
-            "name": "ftrack",
-            "actions": actions
-        }
-    ]
