+++ projects/framework-flame/source/ftrack_framework_flame/__init__.py
@@ -4,7 +4,7 @@ import logging
 import os
 import traceback
 import platform
-
+import functools
 import ftrack_api
 
 from ftrack_framework_core.host import Host
@@ -119,26 +119,69 @@ def bootstrap_integration(framework_extensions_path):
     Host(event_manager, registry=registry_instance)
     client_instance = Client(event_manager, registry=registry_instance)
 
-    # Init tools
-    # dcc_config = registry_instance.get_one(
-    #     name='framework-flame', extension_type='dcc_config'
-    # )['extension']
-    #
-    # logger.debug(f'Read DCC config: {dcc_config}')
-
-    # Create ftrack menu
-
-    # # Register tools into ftrack menu
-    # for tool in dcc_config['tools']:
-    #     pass
-
     return client_instance, registry_instance
 
+def scope_node(show_on, selection):
+    for item in selection:
+        if isinstance(item, show_on):
+            return True
 
-# # Find and read DCC config
-# try:
-#     bootstrap_integration(get_extensions_path_from_environment())
-# except:
-#     # Make sure any exception that happens are logged as there is most likely no console
-#     logger.error(traceback.format_exc())
-#     raise
+    return False
+
+
+def get_ftrack_menu(show_on=None):
+    show_on = show_on or ()
+    client_instance, registry_instance = bootstrap_integration(get_extensions_path_from_environment())
+
+    dcc_config = registry_instance.get_one(
+        name='framework-flame', extension_type='dcc_config'
+    )['extension']
+
+    logger.debug(f'Read DCC config: {dcc_config}')
+
+    # Create ftrack menu
+    actions = []
+    # Register tools into ftrack menu
+    for tool in dcc_config['tools']:
+        run_on = tool.get("run_on")
+        on_menu = tool.get("menu", True)
+        if on_menu:
+            new_action = {
+                'name': tool['label'],
+                'execute': functools.partial(
+                    on_run_tool_callback,
+                    client_instance,
+                    tool.get('name'),
+                    tool.get('dialog_name'),
+                    tool['options'],
+                ),
+                'minimumVersion': '2023',
+                'isVisible': functools.partial(
+                    scope_node, show_on
+                ),
+            }
+            actions.append(new_action)
+
+        if run_on:
+            if run_on == "startup":
+                # Execute startup tool-configs
+                on_run_tool_callback(
+                    client_instance,
+                    tool.get('name'),
+                    tool.get('dialog_name'),
+                    tool['options'],
+                )
+            else:
+                logger.error(
+                    f"Unsupported run_on value: {run_on} tool section of the "
+                    f"tool {tool.get('name')} on the tool config file: "
+                    f"{dcc_config['name']}. \n Currently supported values:"
+                    f" [startup]"
+                )
+
+    return [
+        {
+            "name": "ftrack",
+            "actions": actions
+        }
+    ]
