+++ projects/framework-flame/source/ftrack_framework_flame/utils/__init__.py
@@ -3,12 +3,6 @@
 import threading
 from functools import wraps
 
-# Dock widget in Flame
-def dock_flame_right(widget):
-    '''Dock *widget* to the right side of Flame.'''
-    pass
-
-
 def run_in_main_thread(f):
     '''Make sure a function runs in the main Flame thread.'''
